pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "bysykkel"
description = "Oslo Bysykkel Sanntidsklient"
readme = "README.md"
requires-python = ">=3.7"
keywords = []
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "httpx~=0.23.1",
    "typer~=0.7.0",
    "fastapi~=0.88.0",
    "pydantic~=1.10.2",
    "rich~=12.6.0",
    "python-dotenv~=0.21.0",
    "uvicorn~=0.20.0",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = ["pytest~=7.2.0", "pytest-asyncio~=0.20.3",  "respx~=0.20.1"]

[project.scripts]
bysykkel = "bysykkel.cli.main:app"
